% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_binaries_by_os}
\alias{get_binaries_by_os}
\title{Get binaries for an R Universe build by OS, R version, and arch}
\usage{
get_binaries_by_os(
  uni,
  uni_os_branch,
  r_version,
  bioc_version,
  os,
  macosx_name = NULL,
  arch = NULL
)
}
\arguments{
\item{uni}{character universe name}

\item{uni_os_branch}{character release or devel}

\item{r_version}{character}

\item{bioc_version}{character}

\item{os}{character}

\item{macosx_name}{big-sur or sonoma}

\item{arch}{character x86_64 or arm64}
}
\value{
data.frame packages in a universe
}
\description{
Get binaries for an R Universe build by OS, R version, and arch
}
\examples{
get_binaries_by_os("bioc", "devel", "4.6.0", "3.23", "windows")

}
