% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_binary_os}
\alias{get_binary_os}
\title{Get OS abbreviation}
\usage{
get_binary_os(os)
}
\arguments{
\item{os}{character}
}
\value{
character abbreviation
}
\description{
Get OS abbreviation
}
\examples{
get_binary_os("windows")

}
