% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_candidate_pkgs_for_bioc_version}
\alias{get_candidate_pkgs_for_bioc_version}
\title{Wrapper for get_uni_pkgs, pass information for correct R Universe associated
with a Bioductor version for a specific OS}
\usage{
get_candidate_pkgs_for_bioc_version(os, version)
}
\arguments{
\item{os}{character}

\item{version}{character Bioconductor version}
}
\description{
Wrapper for get_uni_pkgs, pass information for correct R Universe associated
with a Bioductor version for a specific OS
}
