% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_candidates}
\alias{get_candidates}
\title{Get packages matching criteria}
\usage{
get_candidates(
  pkgs,
  commit = FALSE,
  version = FALSE,
  check = c("NOTE", "WARNING", "OK")
)
}
\arguments{
\item{commit}{logical Check commit hash in RU and BBS?}

\item{version}{logical Check package version in RU and BBS?}

\item{check}{character vector of acceptable R CMD check statuses}
}
\value{
data.frame of filtered candidate packages
}
\description{
Get packages matching criteria
}
\examples{
comparable_pkgs <- get_comparable_pkgs("bioc", "devel", "4.6.0", "windows",
                                       "3.23")
pkgs <- get_candiates(comparable_pkgs, commit = TRUE, version = TRUE)

}
