% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_comparable_pkgs}
\alias{get_comparable_pkgs}
\title{Get information about comparable builds in R Universe and the BBS}
\usage{
get_comparable_pkgs(
  uni,
  uni_os_branch,
  r_version,
  bioc_version,
  os,
  macosx_name = NULL,
  arch = NULL,
  verbose = FALSE
)
}
\arguments{
\item{uni}{character universe name}

\item{uni_os_branch}{character release or devel}

\item{r_version}{character}

\item{bioc_version}{character}

\item{os}{character}

\item{macosx_name}{big-sur or sonoma}

\item{arch}{character x86_64 or arm64}
}
\value{
character abbreviation
}
\description{
Collects information from the BBS for a Bioconductor version and finds
information about corresponding package versions in R Universe
}
\examples{
get_comparable_pkgs("bioc", "devel", "4.6.0", "3.23", "windows")

}
