% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_macosx_subpath}
\alias{get_macosx_subpath}
\title{Get macosx repo subpath by R version}
\usage{
get_macosx_subpath(r_version, arch)
}
\arguments{
\item{r_version}{R X.Y character}

\item{arch}{character x86_64 or arm64}
}
\value{
character
}
\description{
Get macosx repo subpath by R version
}
\examples{
get_macosx_subpath("4.6", "arm64")


}
