% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subpath}
\alias{get_subpath}
\title{Get repo subpath by R version}
\usage{
get_subpath(os, arch, r_xy, cran_subpath = FALSE)
}
\arguments{
\item{os}{character OS name, e.g. "macosx", "windows", "linux"}

\item{arch}{character x86_64 or arm64}

\item{r_xy}{R X.Y character}

\item{cran_subpath}{(default FALSE) uses CRAN style subpaths}
}
\value{
character
}
\description{
Get repo subpath by R version
}
\examples{
get_subpath("macosx", "arm64", "4.6")
get_subpath(c("macosx", "windows"), c("arm64", "x86_64"), c("4.6", "4.5"))

}
