% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uni_for_bioc_version}
\alias{get_uni_for_bioc_version}
\title{Get information about R Universe building a Bioconductor version}
\usage{
get_uni_for_bioc_version(version)
}
\arguments{
\item{version}{character Bioconductor version}
}
\description{
Get information about R Universe building a Bioconductor version
}
