% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_greater_than}
\alias{is_greater_than}
\title{Check if the first version is greater the second}
\usage{
is_greater_than(version1, version2)
}
\description{
Check if the first version is greater the second
}
\examples{
is_greater_than("0.10.3", "1.0.20")

}
