% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_xy_ver}
\alias{r_xy_ver}
\title{Get x.y version of R given x.y.z}
\usage{
r_xy_ver(version)
}
\arguments{
\item{version}{character}
}
\description{
Get x.y version of R given x.y.z
}
\examples{
r_xy_ver("4.5.2")

}
