% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_binaries}
\alias{remove_binaries}
\title{Remove old binaries if a new binary exists}
\usage{
remove_binaries(pkgs, repo_root, os, test = TRUE)
}
\arguments{
\item{pkgs}{Data.Frame of packages that includes new binaries}

\item{repo_root}{path that includes repo type--e.g.,
"/home/biocpush/PACKAGES/3.22/bioc"}

\item{os}{name, full or abbreviation}

\item{test}{logical (default TRUE) don't remove, only print packages marked
for removal}
}
\value{
An invisible \code{NULL}
}
\description{
Remove old binaries if a new binary exists
}
\examples{
allpkgs <- get_uni_pkgs("bioc", "devel", "3.23", "windows")
pkgs <- get_candidates(allpkgs)
remove_binaries(pkgs, "/home/biocpush/PACKAGES/3.22/bioc", "windows")

}
