% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_old_binaries}
\alias{remove_old_binaries}
\title{Remove old binaries if a new binary exists}
\usage{
remove_old_binaries(
  repo_root,
  r_version,
  os,
  arch = NA_character_,
  test = TRUE
)
}
\arguments{
\item{repo_root}{path that includes repo type--e.g.,
"/home/biocpush/PACKAGES/3.22/bioc"}

\item{r_version}{R x.y.z version}

\item{os}{name, full or abbreviation}

\item{arch}{x86_64 or arm64}

\item{test}{logical (default TRUE) don't remove, only print packages marked
for removal}
}
\value{
vector of the full path of binaries removed
}
\description{
Remove old binaries if a new binary exists
}
\examples{
\dontrun{
bu <- uni_for_bioc("devel")
repo_root <- paste0("/home/biocpush/PACKAGES/", bu$bioc_version, "/bioc")
remove_old_binaries(repo_root, bu$bioc_version, "windows", "x86_64")
}

}
