% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.expand_jobs}
\alias{.expand_jobs}
\title{get_uni_df helper to expand \verb{_jobs}}
\usage{
.expand_jobs(df)
}
\arguments{
\item{df}{data.frame unprocessed data from R Universe API}
}
\value{
data.frame
}
\description{
Adds additional columns to job_r_xy, job_type, job_arch, job_os
}
\examples{
raw_df <- get_raw_uni_df("bioc")
df <- .expand_jobs(raw_df)

}
