% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_arch}
\alias{filter_by_arch}
\title{Filter by arch if it has arch-specific and universal binaries}
\usage{
filter_by_arch(df, os, arch)
}
\arguments{
\item{df}{data.frame from get_jobs()}

\item{os}{character}

\item{arch}{character}
}
\value{
data.frame
}
\description{
Some builds can have arch-specific and universal binaries. This
function filters specifically for data observed in R Universe, which could
change in the future.
\itemize{
\item mac universal binaries: job_arch == 'arm64' & !is.na(binaries_arch)
\item linux default binaries: job_arch == 'x86_64' & !is.na(binaries_arch)
}
}
\examples{
bu <- uni_for_bioc("devel")
df <- get_jobs(bu)
filter_by_os_arch(df, "macosx", "arm64")

}
