% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_arch}
\alias{filter_by_arch}
\title{Filter by arch if it has arch-specific and universal binaries}
\usage{
filter_by_arch(df, os, arch = NA)
}
\arguments{
\item{df}{data.frame from get_jobs()}

\item{os}{character}

\item{arch}{character}
}
\value{
data.frame
}
\description{
Some builds can have arch-specific and universal binaries. This
function filters specifically for data observed in R Universe, which could
change in the future.
\itemize{
\item mac universal binaries: \verb{_jobs_arch} == 'arm64' & !is.na(\verb{_binaries_arch})
\item linux universal binaries: \verb{_jobs_arch} == 'x86_64' & !is.na(\verb{_binaries_arch})
}
}
\examples{
bu <- get_uni_for_bioc_version("devel")
df <- get_jobs(bu$ru_uni, bu$bioc_branch, bu$r_version, bu$bioc_version)
filter_by_arch(df, "macosx", "arm64")

}
