% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arch}
\alias{get_arch}
\title{Get the arch}
\usage{
get_arch(binaries_arch, jobs_arch)
}
\arguments{
\item{binaries_arch}{_binaries_arch}

\item{jobs_arch}{_jobs_arch}
}
\value{
character
}
\description{
Get the arch
}
\examples{
get_arch(NA, "arm64")
}
