% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arch}
\alias{get_arch}
\title{Get the arch}
\usage{
get_arch(binaries_arch, job_arch)
}
\arguments{
\item{binaries_arch}{binaries_arch}

\item{job_arch}{job_arch}
}
\value{
character
}
\description{
Get the arch
}
\examples{
get_arch(binaries_arch, "arm64")
}
