% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_candidates}
\alias{get_candidates}
\title{Get candidate packages in R Universe based on criteria}
\usage{
get_candidates(
  bu,
  os,
  arch = NA_character_,
  vignettes = TRUE,
  commit = FALSE,
  unsupported_platforms = TRUE,
  check = c("NOTE", "WARNING", "OK")
)
}
\arguments{
\item{bu}{result of uni_for_bioc}

\item{os}{character}

\item{arch}{character (default NA_character_) x86_64 or arm64}

\item{vignettes}{logical (default TRUE) Check status of source / vignettes}

\item{commit}{logical (default FALSE) Check commit hash in RU and BBS?}

\item{unsupported_platforms}{logical (default TRUE) Check for unsupported
platforms}
}
\value{
data.frame of filtered candidate packages
}
\description{
Filters packages by same commit in  BBS, vignettes passing, or
check status values.
}
\examples{
bu <- uni_for_bioc("devel")
candidates <- get_candidates(bu, "windows", commit = TRUE)

}
