% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_candidates}
\alias{get_candidates}
\title{Get candidate packages in R Universe based on criteria}
\usage{
get_candidates(
  uni,
  uni_os_branch,
  r_version,
  bioc_version,
  os,
  arch = NA,
  vignettes = TRUE,
  commit = FALSE,
  check = c("NOTE", "WARNING", "OK")
)
}
\arguments{
\item{uni}{character universe name}

\item{uni_os_branch}{character release or devel}

\item{r_version}{character}

\item{bioc_version}{character}

\item{os}{character}

\item{arch}{character x86_64 or arm64}

\item{vignettes}{logical (default TRUE) Check status of source / vignettes}

\item{commit}{logical (default FALSE) Check commit hash in RU and BBS?}

\item{check}{character vector of acceptable R CMD check statuses}

\item{available_pkgs}{data.frame of R Universe packages}
}
\value{
data.frame of filtered candidate packages
}
\description{
Filters packages by same commit in  BBS, vignettes passing, or
check status values.
}
\examples{
candidates <- get_candidates("bioc", "devel", "4.6.0", "3.23", "windows",
                             commit = TRUE)

}
