% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_candidates}
\alias{get_candidates}
\title{Get packages matching criteria}
\usage{
get_candidates(pkgs, commit = FALSE, check = c("NOTE", "WARNING", "OK"))
}
\arguments{
\item{commit}{logical Check commit hash in RU and BBS?}

\item{check}{character vector of acceptable R CMD check statuses}

\item{data.frame}{of R Universe packages}
}
\value{
data.frame of filtered candidate packages
}
\description{
Get packages matching criteria
}
\examples{
comparable_pkgs <- get_comparable_pkgs("bioc", "devel", "4.6.0", "windows",
                                       "3.23")
pkgs <- get_candidates(comparable_pkgs, commit = TRUE)

}
