% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_jobs}
\alias{get_jobs}
\title{Get jobs for an R Universe build}
\usage{
get_jobs(uni, uni_os_branch, r_version, bioc_version)
}
\arguments{
\item{uni}{character universe name}

\item{uni_os_branch}{character release or devel}

\item{r_version}{character}

\item{bioc_version}{character}
}
\value{
data.frame packages in a universe
}
\description{
Adds Artifact and JobUrl
}
\examples{
bu <- get_uni_for_bioc_version("devel")
get_jobs(bu$ru_uni, bu$bioc_branch, bu$r_version, bu$bioc_version)

}
