% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_jobs}
\alias{get_jobs}
\title{Get jobs for an R Universe build}
\usage{
get_jobs(bu)
}
\arguments{
\item{bu}{result of uni_for_bioc}
}
\value{
data.frame packages in a universe
}
\description{
Adds Artifact and JobUrl
}
\examples{
bu <- uni_for_bioc("devel")
get_jobs(bu)

}
