% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_jobs}
\alias{get_jobs}
\title{Get jobs for an R Universe build}
\usage{
get_jobs(universe_df, r_version, universe)
}
\arguments{
\item{universe_df}{data.frame output of get_uni_df}

\item{bu}{result of uni_for_bioc}
}
\value{
data.frame packages in a universe
}
\description{
Adds artifact and job_url
}
\examples{
bu <- uni_for_bioc("devel")
raw_universe_df <- get_raw_uni_df(bu$universe)
universe_df <- get_uni_df(raw_universe_df)
get_jobs(universe_df, bu$r_version, bu$universe)

}
