% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raw_uni_df}
\alias{get_raw_uni_df}
\title{Get packages in an R Universe}
\usage{
get_raw_uni_df(uni)
}
\arguments{
\item{uni}{character universe name}
}
\value{
data.frame packages in a universe
}
\description{
Get packages in an R Universe
}
\examples{
get_raw_uni_df("bioc")

}
