% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_repository_path}
\alias{get_repository_path}
\title{Get repo path}
\usage{
get_repository_path(
  repo_root,
  r_version,
  os,
  arch = NA_character_,
  cran_subpath = FALSE
)
}
\arguments{
\item{r_version}{character}

\item{os}{character}

\item{arch}{character x86_64 or arm64}

\item{cran_subpath}{(default FALSE) uses CRAN style subpaths}

\item{bioc_version}{character}
}
\value{
character
}
\description{
Get repo path
}
\details{
R Universe has adopted \if{html}{\out{<os name>}}-\if{html}{\out{<arch>}} style path for linux and
and macosx, \if{html}{\out{<compiler>}}-\if{html}{\out{<arch>}} for windows. CRAN has not adopted these
conventions for linux and windows, so Bioconductor has not yet adopted them.
The \code{cran_subpath} parameter should be TRUE to not use R Universe subpath
for windows.
}
\examples{
repo_root <- "/home/biocpush/PACKAGES/3.22/bioc"
get_repository_path(repo_root, "4.6.0", "windows", "arm64")

}
