% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_repository_path}
\alias{get_repository_path}
\title{Get repo path}
\usage{
get_repository_path(
  repo_root,
  r_version,
  os,
  macosx_name = NA_character_,
  arch = NA_character_
)
}
\arguments{
\item{r_version}{character}

\item{os}{character}

\item{macosx_name}{big-sur or sonoma}

\item{arch}{character x86_64 or arm64}

\item{bioc_version}{character}
}
\value{
character
}
\description{
Get repo path
}
\examples{
get_repository_path("/home/biocpush/PACKAGES/3.22/bioc", "4.6.0", "windows")

}
