% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uni_df}
\alias{get_uni_df}
\title{Flatten raw universe data.frame by matching R, OS, and arch information from
\verb{_jobs} with \verb{_binaries}}
\usage{
get_uni_df(raw_df)
}
\arguments{
\item{raw_df}{raw data.frame from R Universe API}
}
\value{
data.frame
}
\description{
Adds additional columns to job_r_xy, binary_r_xy, job_type, job_arch,
job_os, binary_os
}
\examples{
raw_df <- get_raw_uni_df("bioc")
get_uni_df(raw_df)

}
