% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uni_df}
\alias{get_uni_df}
\title{Flatten raw universe data.frame by matching R, OS, and arch information from
\verb{_jobs} with \verb{_binaries}}
\usage{
get_uni_df(raw_df)
}
\arguments{
\item{raw_df}{raw data.frame from R Universe API}
}
\value{
data.frame
}
\description{
Adds additional columns to \verb{_jobs_r_xy}, \verb{_binaries_r_xy},
\verb{_jobs_type}, \verb{_jobs_arch}, \verb{_jobs_os_}, \verb{_binaries_os_}
}
\examples{
raw_df <- get_raw_uni_df("bioc", "devel", "4.6.0")
get_uni_df(raw_df)

}
