% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uni_pkgs}
\alias{get_uni_pkgs}
\title{Get information about comparable builds in R Universe and the BBS}
\usage{
get_uni_pkgs(uni, uni_os_branch, r_version, os, bioc_version, arch = NULL)
}
\arguments{
\item{uni}{character universe name}

\item{uni_os_branch}{character release or devel}
}
\value{
character abbreviation
}
\description{
Collects information from the BBS for a Bioconductor version and finds
information about corresponding package versions in R Universe
}
\examples{
get_uni_pkgs("bioc", "devel", "4.6.0", "windows", "3.23")

}
