% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_greater_than}
\alias{is_greater_than}
\title{Check if the version in a file is greater than another}
\usage{
is_greater_than(file1, file2)
}
\description{
Check if the version in a file is greater than another
}
\examples{
is_greater_than("bedbaser_0.10.3.tar.gz", "bedbaser_1.0.20.tar.gz")

}
