% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uni_for_bioc}
\alias{uni_for_bioc}
\title{Get information about R Universe building a Bioconductor version}
\usage{
uni_for_bioc(branch)
}
\arguments{
\item{branch}{character Bioconductor branch: "devel" or "release"}
}
\description{
Get information about R Universe building a Bioconductor version
}
\examples{
bu <- uni_for_bioc("devel")

}
