% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uni_pkg_file}
\alias{uni_pkg_file}
\title{Construct binary file name}
\usage{
uni_pkg_file(pkg, os, version)
}
\arguments{
\item{pkg}{character package name}

\item{os}{character OS name}

\item{version}{character package version}
}
\value{
character file with extension
}
\description{
Construct binary file name
}
\examples{
uni_pkg_file("bedbaser", "windows", "1.0.12")

}
