% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uni_repo_url}
\alias{uni_repo_url}
\title{Get universe URL for an os and R version}
\usage{
uni_repo_url(uni, r_version, os, macosx_name = NA, arch = NA)
}
\arguments{
\item{uni}{character universe name}

\item{r_version}{character}

\item{os}{character}

\item{macosx_name}{big-sur or sonoma}

\item{arch}{character x86_64 or arm64}
}
\value{
character
}
\description{
Get universe URL for an os and R version
}
\examples{
uni_repo_url("bioc", "4.5.3", "windows")

}
