% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uni_repo_url}
\alias{uni_repo_url}
\title{Get universe URL for an os and R version}
\usage{
uni_repo_url(uni, r_version, os, arch = NULL)
}
\description{
Get universe URL for an os and R version
}
\examples{
uni_repo_url("bioc", "4.5.3", "windows"))

}
